Update parser and lexer for function arguments support
The parser and lexer were updated to support function arguments. Tokens and rules were added to account for multiple arguments in function declarations and function calls. Moreover, CMakeLists was updated to require version 3.12 of CMake, up from version 3.10. Testing was improved with the inclusion of new test files that cover the use of function arguments.